Struct build_env::BuildEnv

pub struct BuildEnv { /* fields omitted */ }

Allow retrieval of values pretaining to a build process that may be related to the target and/or host triple.

Methods

impl BuildEnv

pub fn from_env() -> Result<BuildEnv, VarError>

Use environment variables (such as those set by cargo) to determine values for target and host via the environment variables TARGET and HOST.

pub fn new_cross(host: String, target: String) -> BuildEnv

Construct a BuildEnv where the host and target may be different.

pub fn new(trip: String) -> BuildEnv

Construct a BuildEnv where target and host are the same.

pub fn target(&self) -> &str

The target we're supplying values for

pub fn host(&self) -> &str

The host we're supplying values for

pub fn var<K: AsRef<OsStr>>(&self, var_base: K) -> Option<OsString>

Query the environment for a value, trying the most specific first, before querying more general variables.

<var>_<target> - for example, CC_x86_64-unknown-linux-gnu

<var>_<target_with_underscores> - for example, CC_x86_64_unknown_linux_gnu

<build-kind>_<var> - for example, HOST_CC or TARGET_CFLAGS

<var> - a plain CC, AR as above.

pub fn var_str<K: AsRef<OsStr> + Debug + Any>(
&self,
var_base: K
) -> Result<String, VarError<K>>

The same as Self::var(), but converts the return to an OsString and provides a useful error message